chore: split the main file into different submodules
