+++ src/utils.js
@@ -1,0 +1,8 @@
+/**
+ * Generates a random id.
+ *
+ * @return {string}
+ * */
+export function generateId () {
+  return Math.random().toString().slice(2)
+}
